Extract UnifiedDiffView into its own file to reduce changes-tab.tsx complexity
changes-tab.tsx grew to 1419 lines with multiple large components co-located.
Extract UnifiedDiffView and its helpers (refractor setup, language mappings,
diff utilities) into unified-diff-view.tsx (~530 lines), dropping the parent
file to 893 lines.
